refactor: separate form from page
